+++ NEWS.md
@@ -8,6 +8,7 @@ Version 0.12.5 For Dev-C++ 7 Beta
  - fix: wrong executable filename for source files in linux
  - enhancement: console pauser for linux 
  - enhancement: redirect input to program in linux
+ - enhancement: detach pausing console window
 
 Version 0.12.4 For Dev-C++ 7 Beta
  - change: add copyright infos to each source file